create caiso_utils.py
